schemagen: Fix/remove old code and comments
Copy was not working correctly and was not needed any more, so it has
been removed.

Also the big block of comments in schemaFromStruct was from an old
implementation of the function that worked off of inferred behaviour
from the rest of the filter data. Actions have since been moved to an
explicit type in the filter workflow.