schemagen: Update to use TabPrinter, add expandFprint
This is the next step - printing expanders.

This also adds an extra set of state data to the generator, which is
generated when the initial schema is generated, keeping it in sync with
what is being generated here and using the same set of filters to
influence what gets generated for expanders.

Also, converted schemaFprint to use TabPrinter. This makes the code
much easier to read and means we don't have to keep track of the
tab/indent count in the generator itself.